requirements.txt
streamlit
transformers
torch
pytesseract
Pillow
opencv-python
numpy
